# Improve performance of searching the subscriptions admin list table when using WP_Post data structures