--- a/includes/data-stores/class-wcs-subscription-data-store-cpt.php
+++ b/includes/data-stores/class-wcs-subscription-data-store-cpt.php
@@ -534,10 +534,17 @@ class WCS_Subscription_Data_Store_CPT extends WC_Order_Data_Store_CPT implements
 		global $wpdb;
 
 		$subscription_ids = array();
-		$search_fields    = array_map( 'wc_clean', apply_filters( 'woocommerce_shop_subscription_search_fields', array(
-			'_billing_address_index',
-			'_shipping_address_index',
-		) ) );
+
+		$search_fields = array_map(
+			'wc_clean',
+			apply_filters(
+				'woocommerce_shop_subscription_search_fields',
+				[
+					'_billing_address_index',
+					'_shipping_address_index',
+				]
+			)
+		);
 
 		if ( is_numeric( $term ) ) {
 			$subscription_ids[] = absint( $term );